Move RateLimiterConfig to core
